adfinfo: allow showing entries deeper in dir. tree.
So far, adfinfo was able to show only directory entries
in the main directory and in the 1st level of subdirectories
on a volume.
(This is due to using only adfChangeDir() which allows to cross
only one level).

This modification removes this limitation, any entry in the whole
directory tree on a volume can be checked.