examples / common / change_dir: do nothing if dir path is empty.
Although the function was working without this, it is more optimal
to check this at the beginning.

Note that ".", which is a special, reserved name in unix-like OSs
for "the current directory", is a perfectly legal file or directory
name in AmigaDOS... So - the entries named "." (or "..") can be
encountered on Amiga disks(!).

Useful links:
https://stackoverflow.com/questions/1976007/what-characters-are-forbidden-in-windows-and-linux-directory-names