adfDevGetClassBySizeBlocks: assume HD(F) if not a predefined device.
So far, the code was assuming HD(F) device only if the device was
larger than floppies. In reality, there are many HDF files smaller
than standard 880k Amiga floppies. So HDFs are, apparently, used
to store AmigaDOS volumes of any sizes.
That assumption was invalid and the HD(F)s smaller than floppies
were simply not accessible using the library.

This modification fixes the issue, so that ADFlib allows to access
HD(F)s of small sizes.