tests / examples2: improve the script for generating results for msys.
So far, the script was deleting the results for msys (with
the postfix "_msys") and generating new ones. Since at present
there are usually just a few files to update or to create, it is
better to generate only those that are missing.

Also, calling the unix2dos was changed to use the -n option instead
of input/output redirection. Like this, the binary results (eg.
testing unadf -p) are not changed.