tests adfls: fix tests on Windows and MacOS.
Reordering the command line arguments calling adfls.

The tests were passing on Linux, but on Windows and MacOS,
the command-line arguments seems to be interpreted as filenames.
Apparently, getopt on Windows and on MacOS does not accept options
(switches) after the positional arguments (here - after the first
argument, the adf file).

This modification fixes that issue.